Rewrite everything. The script wasn't the most convenient, so it's been rewritten to have a web front end which shows better insight into the codecs and the encoding progress. Errors show as toasts (and will show again when the page is refreshed, in case you missed one).
Previously it'd just try and re-encode with only the Japanese audio track, if there's a tag tracked as such. This changes that and will mux a video for each audio track. Web browsers don't have good support for switching audio tracks in the same video container, and sidestepping it by syncing an audio and video element didn't work reliably between browsers (always desynced in Firefox, worked in Chrome). So to avoid having to do real-time transcoding this just will mux a video file for each audio track combination. This also saves a "[videofilename]-metadata.json" file that will have a list of the exported videos for each source video and some audio stream metadata corresponding to each, so Jimaku Player can look for those metadata files to find unique sets of videos (and not show what look like duplicate videos) and let the user choose which audio stream they want to listen to.

The metadata file also will include the text content for srt/ass format subtitles, so Jimaku Player can offer to use the subtitles included with the original video.