Rewrite transcoding logic with support for DASH video.
Instead of muxing different full copies of the video with each audio combination it will generate separate files for video and each audio stream, then use MP4Box to dash the video and generate MPD files for each stream. The 'safe codecs' are currently limited to a smaller subset as some codecs are ok but some profiles are not supported for Media Source Extensions (the technology used to stream DASH video).